+++ raycast-env.d.ts
@@ -7,7 +7,22 @@
 
 /* eslint-disable @typescript-eslint/ban-types */
 
-type ExtensionPreferences = {}
+type ExtensionPreferences = {
+  /** Markers for Selective Mercy - Show mercy on the humble bullet points (•, -, *). */
+  "spareBullets": boolean,
+  /**  - Grant passage to the numeric order (1. 2. 3.) */
+  "spareNumbers": boolean,
+  /**  - Grant clemency to alphabetic procession (a, b, c) */
+  "spareLetters": boolean,
+  /** Add Collateral Damage - Slice the heads of markdown (#, ##, ###) with extreme prejudice. */
+  "killHeaders": boolean,
+  /**  - Target snarky blockquotes (>) for termination. */
+  "killQuotes": boolean,
+  /**  - Obliterate oppressive text dungeon (```) code blocks. */
+  "killCode": boolean,
+  /**  - Chop horizontal rules (---, ***) into pieces. */
+  "killRules": boolean
+}
 
 /** Preferences accessible in all the extension's commands */
 declare type Preferences = ExtensionPreferences